requirements.txt
aistudio-sdk==0.3.7
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
Authlib==1.6.3
bandit==1.8.6
bce-python-sdk==0.9.45
beautifulsoup4==4.14.2
billiard==4.2.2
black==25.1.0
blessed==1.22.0
celery==5.5.3
celery-types==0.24.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
codespell==2.4.1
colorlog==6.9.0
cron_descriptor==2.0.6
cryptography==45.0.7
cssselect==1.4.0
defusedxml==0.7.1
detect-secrets==1.5.0
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-anymail==13.1
django-browser-reload==1.19.0
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
django-picklefield==3.3
django-q2==1.8.0
django-redis==6.0.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
docutils==0.22
docutils-stubs==0.0.22
dparse==0.6.4
drf-spectacular==0.28.0
drf-yasg==1.21.10
eradicate==2.3.0
filelock==3.19.1
flake8==7.3.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-print==5.0.0
flake8-rst-docstrings==0.3.1
flake8_import_order==0.19.2
flower==2.0.1
fpdf==1.7.2
fsspec==2025.9.0
future==1.0.0
greenlet==3.3.1
gunicorn==20.0.4
h11==0.16.0
hf-xet==1.1.9
honcho==2.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanize==4.15.0
identify==2.6.14
idna==3.10
imagesize==1.4.1
inflection==0.5.1
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mccabe==0.7.0
mdurl==0.1.2
modelscope==1.29.2
mypy==1.17.1
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.3.1
openai==1.107.1
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.3.2
pathspec==0.12.1
pdf2image==1.17.0
pillow==12.0.0
platformdirs==4.4.0
playwright==1.58.0
pre_commit==4.3.0
prettytable==3.16.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
prosemirror==0.6.0
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pydocstyle==6.3.0
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pypdfium2==4.30.0
pyperclip==1.9.0
pytesseract==0.3.13
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==7.1.0
referencing==0.36.2
regex==2025.9.1
reportlab==4.0.4
requests==2.32.5
requests-oauthlib==2.0.0
restructuredtext_lint==1.4.0
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
safety==3.6.1
safety-schemas==0.0.14
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.8
sqlparse==0.5.3
stevedore==5.5.0
stripe==14.4.1
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.1
typer==0.17.4
types-python-dateutil==2.9.0.20250822
types-pytz==2025.2.0.20250809
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
